backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
openai==1.3.7
aiofiles==23.2.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3